UXD-1079 pencil hover focus
